gitignore 수정 및 공통 기능 정리
